backend/requirements.txt
pdfplumber
beautifulsoup4
tiktoken
openai
sentence-transformers
faiss-cpu
fastapi
streamlit
pydantic
uvicorn
python-multipart
markdown

